+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -28,12 +28,12 @@ public class FilmController {
     }
 
     @PutMapping("/{id}/like/{userId}")
-    public void addLike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void addLike(@PathVariable("id") Long id, @PathVariable("userId") Long userId) {
         filmService.addLike(id, userId);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
-    public void deleteLike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void deleteLike(@PathVariable("id") Long id, @PathVariable("userId") Long userId) {
         filmService.deleteLike(id, userId);
     }
 
@@ -43,7 +43,7 @@ public class FilmController {
     }
 
     @GetMapping("/{id}")
-    public Film getFilm(@PathVariable("id") Integer filmId) {
+    public Film getFilm(@PathVariable("id") Long filmId) {
         return filmService.getById(filmId);
     }
 
@@ -56,17 +56,17 @@ public class FilmController {
     }
 
     @GetMapping("/common") //?userId={userId}&friendId={friendId}
-    public List<Film> getCommonFilms(@RequestParam("userId") Integer userId, @RequestParam("friendId") Integer id) {
+    public List<Film> getCommonFilms(@RequestParam("userId") Long userId, @RequestParam("friendId") Long id) {
         return filmService.getCommon(userId, id);
     }
 
     @GetMapping("/director/{directorId}")
-    public List<Film> getDirectorsFilms(@PathVariable("directorId") Integer directorId, @RequestParam String sortBy) {
+    public List<Film> getDirectorsFilms(@PathVariable("directorId") Long directorId, @RequestParam String sortBy) {
         return filmService.getDirectorsFilms(directorId, sortBy);
     }
 
     @DeleteMapping("/{filmId}")
-    public void deleteFilm(@PathVariable("filmId") Integer filmId) {
+    public void deleteFilm(@PathVariable("filmId") Long filmId) {
         filmService.delete(filmId);
     }
