+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/review/ReviewDislikeDao.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.dao.review;
 
 public interface ReviewDislikeDao {
-    void add(Integer id, Integer userId);
+    void add(Long id, Long userId);
 
-    void delete(Integer id, Integer userId);
+    void delete(Long id, Long userId);
 }
