+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/review/impl/ReviewDislikeDaoImpl.java
@@ -11,7 +11,7 @@ public class ReviewDislikeDaoImpl implements ReviewDislikeDao {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void add(Integer id, Integer userId) {
+    public void add(Long id, Long userId) {
         String sqlQuery = "INSERT INTO review_dislikes (review_id, user_id) VALUES (?,?)";
         jdbcTemplate.update(sqlQuery, id, userId);
         sqlQuery = "UPDATE reviews SET useful = useful - 1 WHERE review_id = ?";
@@ -19,7 +19,7 @@ public class ReviewDislikeDaoImpl implements ReviewDislikeDao {
     }
 
     @Override
-    public void delete(Integer id, Integer userId) {
+    public void delete(Long id, Long userId) {
         String sqlQuery = "DELETE FROM review_dislikes WHERE review_id = ? AND user_id = ?";
         jdbcTemplate.update(sqlQuery, id, userId);
         sqlQuery = "UPDATE reviews SET useful = useful + 1 WHERE review_id = ?";
